Full rewrite to modularize the manifests.
The main issue with a mono-manifest is that a single package failing to
build breaks the upgrade (unless pinning the package is an option).
Also, there are many packages for which outdated versions (and
especially outdated, deep dependencies) are acceptable such that the
considerble time and space spent rebuilding these packages is avoided.

The prior mono-manifest was intended for installation from a Guix
development repository, which contained both submitted and in progress
fixes and features. The new manifests can be built and installed from a
development repository but are intended to work with the scripts and
commit files to create profiles, shells, and packs off upstream.

The manifests no longer provide updates to packages. It is very
difficult to maintain these updates, and better to submit patches.